+++ OSLC4Net_SDK/OSLC4Net.Core/Model/AllowedValues.cs
@@ -1,18 +1,24 @@
+using System.Diagnostics.CodeAnalysis;
 using OSLC4Net.Core.Attribute;
 
 namespace OSLC4Net.Core.Model;
 
+[Experimental("OSLCEXP001", Message = "This class may change or be removed in future releases after [OslcAllowedValues] attribute support is fully implemented and verified.")]
 [OslcNamespace(OslcConstants.OSLC_CORE_NAMESPACE)]
 [OslcResourceShape(title = "OSLC Allowed Values Resource Shape",
     describes = new[] { OslcConstants.TYPE_ALLOWED_VALUES })]
-internal sealed class AllowedValues
+public sealed class AllowedValues : IResource
 {
-    private readonly List<string> allowedValues = new();
+    private readonly List<string> allowedValues = [];
+
+    public Uri About { get; set; } = null!;
 
     public void AddAllowedValue(string allowedValue)
     {
         allowedValues.Add(allowedValue);
     }
+
+    public Uri GetAbout() => About;
 
     [OslcDescription("Value allowed for a property")]
     [OslcName("allowedValue")]
@@ -23,6 +29,11 @@ internal sealed class AllowedValues
     public string[] GetAllowedValues()
     {
         return allowedValues.ToArray<string>();
+    }
+
+    public void SetAbout(Uri about)
+    {
+        About = about;
     }
 
     public void SetAllowedValues(string[] allowedValues)