+++ OSLC4Net_SDK/OSLC4Net.Core/Model/Occurs.cs
@@ -39,7 +39,7 @@ public static class OccursExtension
 {
     public static string ToString(Occurs occurs)
     {
-        var attributes = (URI[])occurs.GetType().GetField(occurs.ToString())
+        var attributes = (URI[])occurs.GetType().GetField(occurs.ToString())!
             .GetCustomAttributes(typeof(URI), false);
 
         return attributes.Length > 0 ? attributes[0].uri : string.Empty;